+++ src/Foundation.Cms/Pages/CmsSettingsPage.cs
@@ -1,0 +1,80 @@
+﻿using EPiServer.Core;
+using EPiServer.DataAbstraction;
+using EPiServer.DataAnnotations;
+using EPiServer.Framework.DataAnnotations;
+using EPiServer.SpecializedProperties;
+using Foundation.Cms.Blocks;
+using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace Foundation.Cms.Pages
+{
+    [ContentType(DisplayName = "Cms Settings Page", 
+        GUID = "183e3bec-dba2-4751-872c-5a6e3bf147d4", 
+        Description = "The page that store cms site settings",
+        AvailableInEditMode = false)]
+    [AvailableContentTypes(Availability.Specific, 
+        Include = new Type[] { typeof(CmsSettingsPage) }, 
+        ExcludeOn = new Type[] { typeof(FoundationPageData) })]
+    [ImageUrl("~/assets/icons/cms/pages/CMS-icon-page-29.png")]
+    public class CmsSettingsPage : PageData
+    {
+        #region Menu   
+
+        [AllowedTypes(new[] { typeof(MenuItemBlock), typeof(PageData) })]
+        [UIHint("HideContentAreaActionsContainer", PresentationLayer.Edit)]
+        [Display(Name = "Main menu", GroupName = CmsTabNames.Menu, Order = 10)]
+        public virtual ContentArea MainMenu { get; set; }
+
+        [CultureSpecific]
+        [Display(Name = "My account menu (CMS)",
+            Description = "This menu will show if show commerce components in header is false",
+            GroupName = CmsTabNames.Menu,
+            Order = 40)]
+        public virtual LinkItemCollection MyAccountCmsMenu { get; set; }
+
+        #endregion
+
+        #region Footer
+
+        [Display(Name = "Introduction", GroupName = CmsTabNames.Footer, Order = 10)]
+        public virtual string Introduction { get; set; }
+
+        [Display(Name = "Company header", GroupName = CmsTabNames.Footer, Order = 20)]
+        public virtual string CompanyHeader { get; set; }
+
+        [Display(Name = "Company name", GroupName = CmsTabNames.Footer, Order = 25)]
+        public virtual string CompanyName { get; set; }
+
+        [Display(Name = "Comapny address", GroupName = CmsTabNames.Footer, Order = 30)]
+        public virtual string CompanyAddress { get; set; }
+
+        [Display(Name = "Company phone", GroupName = CmsTabNames.Footer, Order = 40)]
+        public virtual string CompanyPhone { get; set; }
+
+        [Display(Name = "Company email", GroupName = CmsTabNames.Footer, Order = 50)]
+        public virtual string CompanyEmail { get; set; }
+
+        [Display(Name = "Links header", GroupName = CmsTabNames.Footer, Order = 60)]
+        public virtual string LinksHeader { get; set; }
+
+        [UIHint("FooterColumnNavigation")]
+        [Display(Name = "Links", GroupName = CmsTabNames.Footer, Order = 70)]
+        public virtual LinkItemCollection Links { get; set; }
+
+        [Display(Name = "Social header", GroupName = CmsTabNames.Footer, Order = 80)]
+        public virtual string SocialHeader { get; set; }
+
+        [Display(Name = "Social links", GroupName = CmsTabNames.Footer, Order = 85)]
+        public virtual LinkItemCollection SocialLinks { get; set; }
+
+        [CultureSpecific]
+        [Display(Name = "Content area", GroupName = CmsTabNames.Footer, Order = 90)]
+        public virtual ContentArea ContentArea { get; set; }
+
+        [Display(Name = "Copyright", GroupName = CmsTabNames.Footer, Order = 130)]
+        public virtual string FooterCopyrightText { get; set; }
+
+        #endregion
+    }
+}
