+++ src/Foundation.Demo/Models/DemoHomePage.cs
@@ -1,14 +1,12 @@
 using EPiServer.Core;
 using EPiServer.DataAbstraction;
 using EPiServer.DataAnnotations;
-using EPiServer.Shell.ObjectEditing;
-using EPiServer.Web;
+using EPiServer.ServiceLocation;
 using Foundation.Cms;
-using Foundation.Cms.EditorDescriptors;
-using Foundation.Commerce;
-using Foundation.Commerce.Models.EditorDescriptors;
+using Foundation.Cms.SiteSettings;
+using Foundation.Cms.ViewModels;
 using Foundation.Commerce.Models.Pages;
-using Foundation.Find.Cms.Facets;
+using Foundation.Demo.ViewModels;
 using Foundation.Find.Cms.Facets.Config;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -21,66 +19,34 @@ namespace Foundation.Demo.Models
         AvailableInEditMode = true,
         GroupName = CmsGroupNames.Content)]
     [ImageUrl("~/assets/icons/cms/pages/CMS-icon-page-02.png")]
-    public class DemoHomePage : CommerceHomePage, IFacetConfiguration
+    public class DemoHomePage : CommerceHomePage
     {
-        #region Header
-
         [CultureSpecific]
-        [UIHint(UIHint.Image)]
-        [Display(Name = "Site logo", GroupName = CmsTabNames.Header, Order = 10)]
-        public virtual ContentReference SiteLogo { get; set; }
-
-        [SelectOne(SelectionFactoryType = typeof(HeaderMenuSelectionFactory))]
-        [Display(Name = "Menu style", GroupName = CmsTabNames.Header, Order = 30)]
-        public virtual string HeaderMenuStyle { get; set; }
-
-        [Display(Name = "Large header menu", GroupName = CmsTabNames.Header, Order = 35)]
-        public virtual bool LargeHeaderMenu { get; set; }
-
-        [Display(Name = "Show commerce header components", GroupName = CmsTabNames.Header, Order = 40)]
-        public virtual bool ShowCommerceHeaderComponents { get; set; }
-
-        [Display(Name = "Sticky header", GroupName = CmsTabNames.Header, Order = 50)]
-        public virtual bool StickyTopHeader { get; set; }
-
-        #endregion
-
-        #region Search Settings
-
-        [SelectOne(SelectionFactoryType = typeof(SearchOptionSelectionFactory))]
-        [Display(Name = "Search option", GroupName = CommerceTabNames.SearchSettings, Order = 50)]
-        public virtual string SearchOption { get; set; }
-
-        [Display(Name = "Show products in search results", GroupName = CommerceTabNames.SearchSettings, Order = 100)]
-        public virtual bool ShowProductSearchResults { get; set; }
-
-        [Display(Name = "Show contents in search results", GroupName = CommerceTabNames.SearchSettings, Order = 150)]
-        public virtual bool ShowContentSearchResults { get; set; }
-
-        [Display(Name = "Include images in contents search results", GroupName = CommerceTabNames.SearchSettings, Order = 200)]
-        public virtual bool IncludeImagesInContentsSearchResults { get; set; }
-
-        [SelectOne(SelectionFactoryType = typeof(CatalogSelectionFactory))]
-        [Display(Name = "Search catalog", GroupName = CommerceTabNames.SearchSettings, Order = 250,
-            Description = "The catalogs that will be returned by search.")]
-        public virtual int SearchCatalog { get; set; }
-
-        [Display(
-          Name = "Search Filters Configuration",
-          Description = "Manage filters to be displayed on Search",
-          GroupName = CommerceTabNames.SearchSettings,
-          Order = 300)]
-        [EditorDescriptor(EditorDescriptorType = typeof(IgnoreCollectionEditorDescriptor<FacetFilterConfigurationItem>))]
-        public virtual IList<FacetFilterConfigurationItem> SearchFiltersConfiguration { get; set; }
-
-        #endregion
-
-        public override void SetDefaultValues(ContentType contentType)
+        [Display(Name = "Settings page", GroupName = SystemTabNames.Settings, Order = 200)]
+        [AllowedTypes(new[] { typeof(DemoSettingsPage) })]
+        [Required]
+        public override ContentReference SettingsPage { get; set; }
+        
+        // Get settings
+        public override CmsSiteSettings SiteSettings
         {
-            base.SetDefaultValues(contentType);
-
-            LargeHeaderMenu = false;
-            SearchCatalog = 0;
+            get
+            {
+                var _siteSettingsProvider = ServiceLocator.Current.GetInstance<ISiteSettingsProvider>();
+                return _siteSettingsProvider.GetSiteSettings<DemoSiteSettings>(this);
+            }
         }
+
+        public virtual ContentReference SiteLogo => (SiteSettings as DemoSiteSettings).SiteLogo;
+        public virtual string HeaderMenuStyle => (SiteSettings as DemoSiteSettings).HeaderMenuStyle;
+        public virtual bool LargeHeaderMenu => (SiteSettings as DemoSiteSettings).LargeHeaderMenu;
+        public virtual bool ShowCommerceHeaderComponents => (SiteSettings as DemoSiteSettings).ShowCommerceHeaderComponents;
+        public virtual string SearchOption => (SiteSettings as DemoSiteSettings).SearchOption;
+        public virtual bool ShowProductSearchResults => (SiteSettings as DemoSiteSettings).ShowProductSearchResults;
+        public virtual bool ShowContentSearchResults => (SiteSettings as DemoSiteSettings).ShowContentSearchResults;
+        public virtual bool IncludeImagesInContentsSearchResults => (SiteSettings as DemoSiteSettings).IncludeImagesInContentsSearchResults;
+        public virtual int SearchCatalog => (SiteSettings as DemoSiteSettings).SearchCatalog;
+        public virtual IList<FacetFilterConfigurationItem> SearchFiltersConfiguration => (SiteSettings as DemoSiteSettings).SearchFiltersConfiguration;
+        public virtual bool StickyTopHeader => (SiteSettings as DemoSiteSettings).StickyTopHeader;
     }
 }
